refactor(transaction): remove unused transaction bloc files and clean up transaction detail screen
