+++ lib/app.dart
@@ -13,7 +13,6 @@ class App extends StatefulWidget {
 }
 
 class _AppState extends State<App> {
-  
   @override
   Widget build(BuildContext context) {
     return ScreenUtilInit(
