+++ lib/feature/transaction_screen/transaction_bloc/transaction_state.dart
